backend/pyproject.toml
[project]
name = "ai-accountant"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pgvector>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]
xero = [
    "xero-python>=4.0.0",
]
ai = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "instructor[anthropic]>=1.4.0",
    "anthropic>=0.40.0",
    "sentence-transformers>=3.0.0",
    "interpret>=0.6.0",
    "rapidfuzz>=3.9.0",
]
docs = [
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
